Moved the launcher for multi-players mode in a new class.
