+++ src/main/java/nl/tudelft/jpacman/Launcher.java
@@ -3,18 +3,17 @@ package nl.tudelft.jpacman;
 import java.awt.event.KeyEvent;
 import java.io.IOException;
 import java.io.InputStream;
-import java.lang.reflect.Array;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Timer;
-import java.util.stream.Collectors;
 
 import nl.tudelft.jpacman.board.BoardFactory;
 import nl.tudelft.jpacman.board.Direction;
 import nl.tudelft.jpacman.game.Game;
 import nl.tudelft.jpacman.game.GameFactory;
-import nl.tudelft.jpacman.game.MultiplayerWithoutPacmanGame;
-import nl.tudelft.jpacman.level.*;
+import nl.tudelft.jpacman.level.Level;
+import nl.tudelft.jpacman.level.LevelFactory;
+import nl.tudelft.jpacman.level.MapParser;
+import nl.tudelft.jpacman.level.Player;
+import nl.tudelft.jpacman.level.PlayerFactory;
 import nl.tudelft.jpacman.npc.ghost.GhostFactory;
 import nl.tudelft.jpacman.sprite.PacManSprites;
 import nl.tudelft.jpacman.ui.Action;
@@ -30,8 +29,8 @@ public class Launcher {
 
 	private static final PacManSprites SPRITE_STORE = new PacManSprites();
 
-	private PacManUI pacManUI;
-	private Game game;
+	protected PacManUI pacManUI;
+	protected Game game;
 
 	/**
 	 * @return The game object this launcher will start when {@link #launch()}
@@ -53,57 +52,19 @@ public class Launcher {
 	}
 
 	/**
-	 * Creates a new game using the level from {@link #makeLevel()}.
-	 * @return A new Game.
-     */
-	public Game makeMultiPlayerGame(int playerCount) {
-		GameFactory gf = getGameFactory();
-		assert (playerCount > 4 || playerCount < 2);
-		Level level = makeMultiPlayerLevel("/boardmulti"+playerCount+".txt");
-		return gf.createMultiPlayerWithoutPacmanGame(playerCount, level);
-	}
-
-	/**
-	 * Creates a new level using resource located at path.
-	 *
-	 * @param path - The path to the map.
-	 * @return A new level built from the resource.
-	 */
-	public Level makeLevel(String path) {
-		MapParser parser = getMapParser();
-		try (InputStream boardStream = Launcher.class
-				.getResourceAsStream(path)) {
-			return parser.parseMap(boardStream);
-		} catch (IOException e) {
-			throw new PacmanConfigurationException("Unable to create level.", e);
-		}
-	}
-
-	/**
-	 * Creates a new level using resource located at path.
-	 *
-	 * @param path - The path to the map.
-	 * @return A new level built from the resource for multi-player games.
-	 */
-	public Level makeMultiPlayerLevel(String path) {
-		MapParser parser = getMultiplayerMapParser();
-		parser.setCollisionMap(new MultiPlayerCollisions());
-		try (InputStream boardStream = Launcher.class
-				.getResourceAsStream(path)) {
-			return parser.parseMap(boardStream);
-		} catch (IOException e) {
-			throw new PacmanConfigurationException("Unable to create level.", e);
-		}
-	}
-
-	/**
 	 * Creates a new level. By default this method will use the map parser to
 	 * parse the default board stored in the <code>board.txt</code> resource.
-	 *
+	 * 
 	 * @return A new level.
 	 */
 	public Level makeLevel() {
-		return makeLevel("board.txt");
+		MapParser parser = getMapParser();
+		try (InputStream boardStream = Launcher.class
+				.getResourceAsStream("/board.txt")) {
+			return parser.parseMap(boardStream);
+		} catch (IOException e) {
+			throw new PacmanConfigurationException("Unable to create level.", e);
+		}
 	}
 
 	/**
@@ -112,14 +73,6 @@ public class Launcher {
 	 */
 	protected MapParser getMapParser() {
 		return new MapParser(getLevelFactory(), getBoardFactory());
-	}
-
-	/**
-	 * @return A new map parser for multiplayer using the factories from
-	 *         {@link #getLevelFactory()} and {@link #getBoardFactory()}.
-	 */
-	protected MultiPlayerMapParser getMultiplayerMapParser(){
-		return new MultiPlayerMapParser(getLevelFactory(), getBoardFactory());
 	}
 
 	/**
@@ -206,52 +159,6 @@ public class Launcher {
 
 	}
 
-	/**
-	 * Initialize keys for a multiplayer game.
-	 * Ghost1: Arrows
-	 * Ghost2: Z,S,Q,D
-	 * Ghost3: T,G,F,H
-	 * Ghost5: I,K,J,L
-	 * @param builder - The builder needed to set up keys.
-	 * @param game - The game needed to access players.
-     */
-	protected void addMultiPlayerKeys(final PacManUiBuilder builder, final Game game){
-		// There are at least 2 ghosts
-		final Player ghost1 = game.getPlayers().get(0);
-		initializeKeysForPlayer(ghost1, builder, game,
-				new int[]{KeyEvent.VK_UP, KeyEvent.VK_DOWN, KeyEvent.VK_LEFT, KeyEvent.VK_RIGHT});
-		final Player ghost2 = game.getPlayers().get(1);
-		initializeKeysForPlayer(ghost2, builder, game,
-				new int[]{KeyEvent.VK_Z, KeyEvent.VK_S, KeyEvent.VK_Q, KeyEvent.VK_D});
-
-		if(game.getPlayers().size() >= 3){
-			final Player ghost3 = game.getPlayers().get(2);
-			initializeKeysForPlayer(ghost3, builder, game,
-					new int[]{KeyEvent.VK_T, KeyEvent.VK_G, KeyEvent.VK_F, KeyEvent.VK_H});
-		}
-		if(game.getPlayers().size() == 4){
-			final Player ghost4 = game.getPlayers().get(3);
-			initializeKeysForPlayer(ghost4, builder, game,
-					new int[]{KeyEvent.VK_I, KeyEvent.VK_K, KeyEvent.VK_J, KeyEvent.VK_L});
-		}
-	}
-
-	/**
-	 * Initialize the key-mapping for the player given as paramter.
-	 * @param ghost - The ghost which needs a key-binding.
-	 * @param builder - The ui builder.
-	 * @param game - The game.
-	 * @param keys - The keys to set, first is north, second is south,
-	 *               third is west and last is east.
-	 */
-	private void initializeKeysForPlayer(final Player ghost, PacManUiBuilder builder, final Game game, int[] keys) {
-		Direction[] directions = { Direction.NORTH, Direction.SOUTH, Direction.WEST, Direction.EAST};
-		for(int i = 0; i<keys.length; i++) {
-			final int j = i;
-			builder.addKey(keys[j], () -> game.move(ghost, directions[j]));
-		}
-	}
-
 	private Player getSinglePlayer(final Game game) {
 		List<Player> players = game.getPlayers();
 		if (players.isEmpty()) {
@@ -273,17 +180,6 @@ public class Launcher {
 	}
 
 	/**
-	 * Creates and starts a Multi-player Wihtout Pacman game.
-	 */
-	public void launchMultiPlayer(){
-		game = makeMultiPlayerGame(2);
-		PacManUiBuilder builder = new PacManUiBuilder().withDefaultButtons();
-		addMultiPlayerKeys(builder, game);
-		pacManUI = builder.build(game);
-		pacManUI.start();
-	}
-
-	/**
 	 * Disposes of the UI. For more information see {@link javax.swing.JFrame#dispose()}.
 	 */
 	public void dispose() {
@@ -299,6 +195,6 @@ public class Launcher {
 	 *             When a resource could not be read.
 	 */
 	public static void main(String[] args) throws IOException {
-		new Launcher().launchMultiPlayer();//.launch();
+		new Launcher().launch();
 	}
 }
